combined yurbo api
