+++ src/app/yurbo/create/Form.tsx
@@ -45,7 +45,7 @@ export default function Form() {
 
   const onSubmit = async (formikValues: Values) => {
     try {
-      const res = await fetch('/api/yurbo/create', {
+      const res = await fetch('/api/yurbo', {
         method: 'POST',
         body: JSON.stringify(formikValues),
       });
